+++ src/database/interface/delete.js
@@ -1,0 +1,8 @@
+const dbConnect = require("../connection");
+
+module.exports = {
+  async deleteRegister(table, field, value) {
+    const response = await dbConnect(table).where(field, '=', value).del();
+    return response;
+  },
+};